Safari-609.1.11.1/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/persisted-user-state-restoration/scroll-restoration-fragment-scrolling-cross-origin-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_equals: should scroll to fragment expected 800 but got 0

Harness Error (TIMEOUT), message = null

TIMEOUT Manual scroll restoration should take precedent over scrolling to fragment in cross origin navigation Test timed out
